+++ boat-common/src/main/kotlin/xyz/srclab/common/test/TestLogger.kt
@@ -4,7 +4,6 @@ import xyz.srclab.common.base.Current
 import xyz.srclab.common.base.Format.Companion.fastFormat
 import xyz.srclab.common.base.Format.Companion.printFormat
 import xyz.srclab.common.base.asAny
-import xyz.srclab.common.reflect.contractSignatureName
 import java.io.PrintStream
 
 interface TestLogger {
@@ -31,11 +30,53 @@ interface TestLogger {
                 val callFrame = Current.callerFrameOrNull(javaClass.name, "log")
                 printStream.println(
                     "%-${CLASS_NAME_MAX_LENGTH}s(%-${LINE_NUMBER_MAX_LENGTH}d): %s".printFormat(
-                        callFrame?.className?.contractSignatureName(CLASS_NAME_MAX_LENGTH),
+                        callFrame?.className?.abridgeName(CLASS_NAME_MAX_LENGTH),
                         callFrame?.lineNumber,
                         message
                     )
                 )
+            }
+
+            /**
+             * Contracts full signature name:
+             * * abc.xyz.Foo -> a.x.Foo
+             *
+             * If given [this] has an illegal format, the result will be undefined.
+             */
+            private fun CharSequence.abridgeName(maxLength: Int = 0): String {
+
+                val split = this.split(".")
+                val wordCount = split.size
+                val tailIndex = split.size - 1
+
+                fun cutWords(split: List<String>, separatorIndex: Int): String {
+                    val buffer = StringBuilder(split[tailIndex])
+                    for (i in (0 until tailIndex).reversed()) {
+                        buffer.insert(0, ".")
+                        if (i < separatorIndex) {
+                            buffer.insert(0, split[i][0])
+                        } else {
+                            buffer.insert(0, split[i])
+                        }
+                    }
+                    return buffer.toString()
+                }
+
+                val minLength = (wordCount - 1) * 2 + split[wordCount - 1].length
+                if (minLength >= maxLength) {
+                    return cutWords(split, tailIndex)
+                }
+                var newLength = minLength
+                for (i in tailIndex - 1 downTo 0) {
+                    newLength += split[i].length - 1
+                    if (newLength > maxLength) {
+                        return cutWords(split, i + 1)
+                    }
+                    if (newLength == maxLength) {
+                        return cutWords(split, i)
+                    }
+                }
+                return cutWords(split, 0)
             }
 
             companion object {
